arm/arm64: defconfig: Update configs to use the new CROS_EC options
Recently we refactored the CrOS EC drivers moving part of the code from
the MFD subsystem to the platform chrome subsystem. During this change
we needed to rename some config options, so, update the defconfigs
accordingly.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Acked-by: Krzysztof Kozlowski <krzk@kernel.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Series-to: soc@kernel.org
Series-cc: linux-kernel@vger.kernel.org, kernel@collabora.com, groeck@chromium.org, bleung@chromium.org, lee.jones@linaro.org, gwendal@chromium.org
Commit-notes:
Dear all,

This is basically a resend of [1] in order to get patch into the arm-soc
patchwork and can be merged independently of the series. The patch was
originally sent as part of the these series [2] but as defconfig changes
often cause merge conflicts the maintainers prefer to have this merged
through the arm-soc tree. My bad was not including the soc ML from the
begining, so sorry about that.

Thanks,
 Enric

[1] https://lkml.org/lkml/2019/8/23/518
[2] https://lkml.org/lkml/2019/8/23/475
END